Move `getCurrentBranchFromFeed`
